Update afterHomework check
