+++ Homeworks/src/main/java/io/codelex/classesandobjects/practice/BankAccount.java
@@ -35,13 +35,4 @@ public class BankAccount {
         String minusSign = (balance < 0) ? "-" : "";
         return name + ", " + minusSign + "$" + formatBalance;
     }
-
-    public static void main(String[] args) {
-        BankAccount bensen = new BankAccount("Benson", 17.25);
-        BankAccount minusBensen = new BankAccount("Benson", -17.50);
-        System.out.println(bensen.toString());
-        System.out.println(bensen.toString2());
-        System.out.println(minusBensen.toString2());
-        System.out.println(minusBensen.toString());
-    }
 }
