+++ Homeworks/src/main/java/io/codelex/classesandobjects/practice/Exercise4/Movie.java
@@ -30,10 +30,12 @@ public class Movie {
         return pgMovies.toArray(new Movie[0]);
     }
 
-    public static void main(String[] args) {
-        Movie movie1 = new Movie("Casino Royale", "Eon Productions", "PG-13");
-        Movie movie2 = new Movie("Glass", "Buena Vista International", "PG-13");
-        Movie movie3 = new Movie("Spider-Man: Into the Spider-Verse", "Columbia Pictures", "PG");
-
+    @Override
+    public String toString() {
+        return "Movie{" +
+                "title='" + title + '\'' +
+                ", studio='" + studio + '\'' +
+                ", rating='" + rating + '\'' +
+                '}';
     }
 }
