+++ Homeworks/src/main/java/io/codelex/classesandobjects/practice/Exercise4/MovieTest.java
@@ -1,0 +1,15 @@
+package io.codelex.classesandobjects.practice.Exercise4;
+
+public class MovieTest {
+
+    public static void main(String[] args) {
+        Movie movie1 = new Movie("Casino Royale", "Eon Productions", "PG-13");
+        Movie movie2 = new Movie("Glass", "Buena Vista International", "PG-13");
+        Movie movie3 = new Movie("Spider-Man: Into the Spider-Verse", "Columbia Pictures", "PG");
+
+        System.out.println(movie1);
+        System.out.println(movie2);
+        System.out.println(movie3);
+
+    }
+}
